requirements.txt
google-generativeai
streamlit
python-dotenv~=0.21.0